Moving GuestData and EditSelection interfaces to GlobalState.ts - where they are best located
